+++ R/0_imports.R
@@ -32,7 +32,8 @@
 # ------------------------------------------------------------------------------
 
 utils::globalVariables(
-  c("engine", "name", "func", "parsnip", "call_name", ".step", "call_info",
+  c(
+    "engine", "name", "func", "parsnip", "call_name", ".step", "call_info",
     "component", "component_id", "id", "control", ".pred", ".metric",
     ".estimator", ".estimate", "n", "note", "object", "splits", "grid",
     "resamples", ".iter", "mean", ".submodels", "metrics", "data", ".mean",
@@ -41,14 +42,22 @@ utils::globalVariables(
     ".extracts", ".metrics", "value", ".notes", ".loss", ".bound",
     ".column", ".totals", ".value", "direction", ".config", "Freq", "Prediction",
     "Truth", ".seed", ".order", ".iter_model", ".iter_preprocessor",
-    ".iter_config", ".msg_model", "# resamples", "seed", "pre", "type")
+    ".iter_config", ".msg_model", "# resamples", "seed", "pre", "type"
   )
+)
 
 # ------------------------------------------------------------------------------
 
 tidyr_new_interface <- function() {
   utils::packageVersion("tidyr") > "0.8.99"
 }
+
+release_bullets <- function() {
+  c(
+    "Update dependencies with `devtools::install_dev_deps()` and update the test objects via `R CMD BATCH --vanilla inst/test_objects.R`."
+  )
+}
+
 
 # ------------------------------------------------------------------------------
 ## function is called in .onLoad() in zzz.R