+++ tests/testthat/helper-s3.R
@@ -12,9 +12,9 @@ expect_s3_class_bare_tibble <- function(x) {
 
 load_helper_tune_results <- function() {
   list(
-    resample = readRDS(test_path("lm_resamples.rds")),
-    tune = readRDS(test_path("knn_results.rds")),
-    bayes = readRDS(test_path("lm_bayes.rds"))
+    resample = readRDS(test_path("data", "lm_resamples.rds")),
+    tune = readRDS(test_path("data", "knn_results.rds")),
+    bayes = readRDS(test_path("data", "lm_bayes.rds"))
   )
 }