+++ tests/testthat/test-engine-parameters.R
@@ -1,16 +1,9 @@
-context("engine-specific tuning paramters")
-
-library(parsnip)
-library(dials)
-
-## -----------------------------------------------------------------------------
-
-test_that('check for finalization with engine parameters', {
-  pset_1 <- parameters(mtry(), penalty(), mixture())
+test_that("check for finalization with engine parameters", {
+  pset_1 <- parameters(dials::mtry(), dials::penalty(), dials::mixture())
   pset_2 <- pset_1
   pset_2$object[[3]] <- NA
 
-  pset_3 <- parameters(mtry(1:2), penalty(), mixture())
+  pset_3 <- parameters(dials::mtry(1:2), dials::penalty(), dials::mixture())
   pset_4 <- pset_3
   pset_4$object[[3]] <- NA
 
@@ -29,21 +22,23 @@ test_that('check for finalization with engine parameters', {
 
 ## -----------------------------------------------------------------------------
 
-test_that('tuning with engine parameters with dials objects', {
+test_that("tuning with engine parameters with dials objects", {
   skip_if_not_installed("randomForest")
   skip_if(utils::packageVersion("dials") <= "0.0.7")
 
   rf_mod <-
-    rand_forest(min_n = tune()) %>%
-    set_engine("randomForest", maxnodes = tune()) %>%
-    set_mode("regression")
+    parsnip::rand_forest(min_n = tune()) %>%
+    parsnip::set_engine("randomForest", maxnodes = tune()) %>%
+    parsnip::set_mode("regression")
 
   set.seed(192)
   rs <- rsample::vfold_cv(mtcars)
 
   set.seed(19828)
   expect_error(
-    rf_tune <- rf_mod %>% tune_grid(mpg ~ ., resamples = rs, grid = 3),
+    suppressMessages(
+      rf_tune <- rf_mod %>% tune_grid(mpg ~ ., resamples = rs, grid = 3)
+    ),
     regex = NA
   )
   expect_error(
@@ -53,7 +48,9 @@ test_that('tuning with engine parameters with dials objects', {
 
   set.seed(283)
   expect_error(
-    rf_search <- rf_mod %>% tune_bayes(mpg ~ ., resamples = rs, initial = 3, iter = 2),
+    suppressMessages(
+      rf_search <- rf_mod %>% tune_bayes(mpg ~ ., resamples = rs, initial = 3, iter = 2)
+    ),
     regex = NA
   )
   expect_error(
@@ -64,49 +61,48 @@ test_that('tuning with engine parameters with dials objects', {
 
 ## -----------------------------------------------------------------------------
 
-test_that('tuning with engine parameters without dials objects', {
+test_that("tuning with engine parameters without dials objects", {
   skip_if_not_installed("randomForest")
   skip_if(utils::packageVersion("dials") <= "0.0.7")
 
   ## ---------------------------------------------------------------------------
 
   rf_mod <-
-    rand_forest(min_n = tune()) %>%
-    set_engine("randomForest", corr.bias = tune()) %>%
-    set_mode("regression")
+    parsnip::rand_forest(min_n = tune()) %>%
+    parsnip::set_engine("randomForest", corr.bias = tune()) %>%
+    parsnip::set_mode("regression")
 
   grid <-
-    data.frame(min_n = c(5, 10, 5, 10),
-               corr.bias = c(TRUE, TRUE, FALSE, FALSE))
+    data.frame(
+      min_n = c(5, 10, 5, 10),
+      corr.bias = c(TRUE, TRUE, FALSE, FALSE)
+    )
 
   set.seed(192)
   rs <- rsample::vfold_cv(mtcars)
 
   ## ---------------------------------------------------------------------------
 
-  expect_error(
-    rf_tune <- rf_mod %>% tune_grid(mpg ~ ., resamples = rs, grid = 3),
-    regex = "missing some parameter objects"
-  )
+  expect_snapshot(error = TRUE, {
+    rf_tune <- rf_mod %>% tune_grid(mpg ~ ., resamples = rs, grid = 3)
+  })
 
   ## ---------------------------------------------------------------------------
 
   expect_error(
-    rf_tune <- rf_mod %>% tune_grid(mpg ~ ., resamples = rs, grid = grid),
+    suppressMessages(
+      rf_tune <- rf_mod %>% tune_grid(mpg ~ ., resamples = rs, grid = grid)
+    ),
     regex = NA
   )
-  expect_error(
-    p <- autoplot(rf_tune),
-    regex = "Some parameters do not have corresponding"
-  )
+  expect_snapshot(error = TRUE, {
+    p <- autoplot(rf_tune)
+  })
 
   ## ---------------------------------------------------------------------------
 
   set.seed(283)
-  expect_error(
-    rf_search <- rf_mod %>% tune_bayes(mpg ~ ., resamples = rs),
-    regex = "missing some parameter objects"
-  )
+  expect_snapshot(error = TRUE, {
+    rf_search <- rf_mod %>% tune_bayes(mpg ~ ., resamples = rs)
+  })
 })
-
-