+++ tests/testthat/test-merge.R
@@ -1,32 +1,30 @@
-context("merge grid values into objects")
+test_that("recipe merges", {
+  data("Chicago", package = "modeldata")
+  spline_rec <-
+    recipes::recipe(ridership ~ ., data = head(Chicago)) %>%
+    recipes::step_date(date) %>%
+    recipes::step_holiday(date) %>%
+    recipes::step_rm(date, dplyr::ends_with("away")) %>%
+    recipes::step_impute_knn(recipes::all_predictors(), neighbors = tune("imputation")) %>%
+    recipes::step_other(recipes::all_nominal(), threshold = tune()) %>%
+    recipes::step_dummy(recipes::all_nominal()) %>%
+    recipes::step_normalize(recipes::all_numeric_predictors()) %>%
+    recipes::step_bs(recipes::all_predictors(), deg_free = tune(), degree = tune())
+  spline_grid <-
+    tibble::tribble(
+      ~imputation, ~threshold, ~deg_free, ~degree,
+      3L,      0.088,       14L,      1L,
+      6L,      0.058,        8L,      1L,
+      8L,      0.051,       14L,      1L,
+      9L,      0.007,       10L,      1L,
+      1L,      0.032,       15L,      2L,
+      8L,      0.018,        9L,      2L,
+      1L,      0.036,        5L,      1L,
+      10L,        0.1,       10L,      2L,
+      9L,      0.094,       12L,      1L,
+      4L,      0.025,       12L,      1L
+    )
 
-# ------------------------------------------------------------------------------
-
-source(test_path("../helper-objects.R"))
-
-# ------------------------------------------------------------------------------
-
-spline_grid <-
-  tibble::tribble(
-    ~imputation, ~threshold, ~deg_free, ~degree,
-    3L,      0.088,       14L,      1L,
-    6L,      0.058,        8L,      1L,
-    8L,      0.051,       14L,      1L,
-    9L,      0.007,       10L,      1L,
-    1L,      0.032,       15L,      2L,
-    8L,      0.018,        9L,      2L,
-    1L,      0.036,        5L,      1L,
-   10L,        0.1,       10L,      2L,
-    9L,      0.094,       12L,      1L,
-    4L,      0.025,       12L,      1L
-  )
-
-bst_grid <- tibble("funky name \n" = 1:4, rules = rep(c(TRUE, FALSE), each = 2))
-
-# ------------------------------------------------------------------------------
-
-test_that('recipe merges', {
-  
   expect_error(
     spline_updated <- merge(spline_rec, spline_grid),
     NA
@@ -46,11 +44,35 @@ test_that('recipe merges', {
       spline_updated$x[[i]]$steps[[8]]$degree, spline_grid$degree[[i]]
     )
   }
-
 })
 
-test_that('partially recipe merge', {
-  
+test_that("partially recipe merge", {
+  data("Chicago", package = "modeldata")
+  spline_rec <-
+    recipes::recipe(ridership ~ ., data = head(Chicago)) %>%
+    recipes::step_date(date) %>%
+    recipes::step_holiday(date) %>%
+    recipes::step_rm(date, dplyr::ends_with("away")) %>%
+    recipes::step_impute_knn(recipes::all_predictors(), neighbors = tune("imputation")) %>%
+    recipes::step_other(recipes::all_nominal(), threshold = tune()) %>%
+    recipes::step_dummy(recipes::all_nominal()) %>%
+    recipes::step_normalize(recipes::all_numeric_predictors()) %>%
+    recipes::step_bs(recipes::all_predictors(), deg_free = tune(), degree = tune())
+  spline_grid <-
+    tibble::tribble(
+      ~imputation, ~threshold, ~deg_free, ~degree,
+      3L,      0.088,       14L,      1L,
+      6L,      0.058,        8L,      1L,
+      8L,      0.051,       14L,      1L,
+      9L,      0.007,       10L,      1L,
+      1L,      0.032,       15L,      2L,
+      8L,      0.018,        9L,      2L,
+      1L,      0.036,        5L,      1L,
+      10L,        0.1,       10L,      2L,
+      9L,      0.094,       12L,      1L,
+      4L,      0.025,       12L,      1L
+    )
+
   expect_error(
     spline_updated <- merge(spline_rec, spline_grid[, -1]),
     NA
@@ -70,11 +92,22 @@ test_that('partially recipe merge', {
       spline_updated$x[[i]]$steps[[8]]$degree, spline_grid$degree[[i]]
     )
   }
-
 })
 
-test_that('umerged recipe merge', {
-  
+test_that("umerged recipe merge", {
+  data("Chicago", package = "modeldata")
+  spline_rec <-
+    recipes::recipe(ridership ~ ., data = head(Chicago)) %>%
+    recipes::step_date(date) %>%
+    recipes::step_holiday(date) %>%
+    recipes::step_rm(date, dplyr::ends_with("away")) %>%
+    recipes::step_impute_knn(recipes::all_predictors(), neighbors = tune("imputation")) %>%
+    recipes::step_other(recipes::all_nominal(), threshold = tune()) %>%
+    recipes::step_dummy(recipes::all_nominal()) %>%
+    recipes::step_normalize(recipes::all_numeric_predictors()) %>%
+    recipes::step_bs(recipes::all_predictors(), deg_free = tune(), degree = tune())
+  bst_grid <- tibble::tibble("funky name \n" = 1:4, rules = rep(c(TRUE, FALSE), each = 2))
+
   expect_error(
     spline_updated <- merge(spline_rec, bst_grid),
     NA
@@ -94,15 +127,18 @@ test_that('umerged recipe merge', {
       spline_updated$x[[i]]$steps[[8]]$degree, tune()
     )
   }
-
 })
 
 # ------------------------------------------------------------------------------
 
 
 
-test_that('model spec merges', {
-  
+test_that("model spec merges", {
+  bst_model <-
+    parsnip::boost_tree(mode = "classification", trees = tune("funky name \n")) %>%
+    parsnip::set_engine("C5.0", rules = tune(), noGlobalPruning = TRUE)
+  bst_grid <- tibble::tibble("funky name \n" = 1:4, rules = rep(c(TRUE, FALSE), each = 2))
+
   expect_error(
     bst_updated <- merge(bst_model, bst_grid),
     NA
@@ -118,11 +154,14 @@ test_that('model spec merges', {
       rlang::as_quosure(bst_grid$rules[[i]], empty_env())
     )
   }
-
 })
 
-test_that('partially model spec merge', {
-  
+test_that("partially model spec merge", {
+  bst_model <-
+    parsnip::boost_tree(mode = "classification", trees = tune("funky name \n")) %>%
+    parsnip::set_engine("C5.0", rules = tune(), noGlobalPruning = TRUE)
+  bst_grid <- tibble::tibble("funky name \n" = 1:4, rules = rep(c(TRUE, FALSE), each = 2))
+
   expect_error(
     bst_updated <- merge(bst_model, bst_grid[, -1]),
     NA
@@ -137,11 +176,14 @@ test_that('partially model spec merge', {
       rlang::as_quosure(bst_grid$rules[[i]], empty_env())
     )
   }
-
 })
 
-test_that('umerged model spec merge', {
-  
+test_that("umerged model spec merge", {
+  bst_model <-
+    parsnip::boost_tree(mode = "classification", trees = tune("funky name \n")) %>%
+    parsnip::set_engine("C5.0", rules = tune(), noGlobalPruning = TRUE)
+  bst_grid <- tibble::tibble("funky name \n" = 1:4, rules = rep(c(TRUE, FALSE), each = 2))
+
   other_grid <- bst_grid
   names(bst_grid) <- letters[1:2]
   expect_error(
@@ -161,6 +203,4 @@ test_that('umerged model spec merge', {
   #     rlang::get_expr(bst_not_updated$x[[i]]$eng_args$rules), tune()
   #   )
   # }
-
 })
-